I noticed that when initializing color you could create a Color value of Color(r: 1, g: 1, b: 1, a: 1) and somehow only get Color(r: 0, g: 0, b: 0, a: 0) with some random numbers sometimes in place of the zeros. With some testing I realized that using Double for Color instead solved the issue. So it seems like the cpp expects doubles and not floats.
Note: this actually seems to cause issues with Vector. So it's possible that it's not as cut and dry and replacing all instances of Float with Double.

+++ Sources/SimpleExtension/Demo.swift
@@ -80,8 +80,8 @@ class SwiftSprite: Sprite2D {
         time_passed += delta
     
         SwiftSprite.lerp (from: 0.1, to: 10, weight: 1)
-        var newPos = Vector2(x: Float (10 + (10 * sin(time_passed * 2.0))),
-                             y: Float (10.0 + (10.0 * cos(time_passed * 1.5))))
+        var newPos = Vector2(x: Double (10 + (10 * sin(time_passed * 2.0))),
+                             y: Double (10.0 + (10.0 * cos(time_passed * 1.5))))
         
         self.position = newPos
     }
